# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CrossSectionalRequest": {
        "properties": {
          "data_source": {
            "default": "yfinance",
            "enum": [
              "yfinance",
              "vnstock",
              "csv"
            ],
            "title": "Data Source",
            "type": "string"
          },
          "date_range": {
            "$ref": "#/components/schemas/DateRange"
          },
          "factor_type": {
            "enum": [
              "moving_average",
              "distance_from_ma",
              "bollinger",
              "donchian",
              "distance_from_peak"
            ],
            "title": "Factor Type",
            "type": "string"
          },
          "ma_type": {
            "default": "sma",
            "enum": [
              "sma",
              "ema",
              "wma"
            ],
            "title": "Ma Type",
            "type": "string"
          },
          "period": {
            "default": 20,
            "title": "Period",
            "type": "integer"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "title": "Symbols",
            "type": "array"
          },
          "threshold": {
            "default": 0.0,
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "symbols",
          "date_range",
          "factor_type"
        ],
        "title": "CrossSectionalRequest",
        "type": "object"
      },
      "CrossSectionalResponse": {
        "properties": {
          "breadth": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Breadth",
            "type": "object"
          },
          "factor_name": {
            "title": "Factor Name",
            "type": "string"
          },
          "pct_above": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Pct Above",
            "type": "object"
          },
          "universe": {
            "items": {
              "type": "string"
            },
            "title": "Universe",
            "type": "array"
          },
          "universe_median": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Universe Median",
            "type": "object"
          }
        },
        "required": [
          "factor_name",
          "universe",
          "breadth",
          "pct_above",
          "universe_median"
        ],
        "title": "CrossSectionalResponse",
        "type": "object"
      },
...
      "DateRange": {
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "DateRange",
        "type": "object"
      },
...
      "FactorAnalysisResponse": {
        "properties": {
          "current_percentile": {
            "title": "Current Percentile",
            "type": "number"
          },
          "current_value": {
            "title": "Current Value",
            "type": "number"
          },
          "factor_name": {
            "title": "Factor Name",
            "type": "string"
          },
          "history_length_days": {
            "title": "History Length Days",
            "type": "integer"
          },
          "percentiles": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Percentiles",
            "type": "object"
          }
        },
        "required": [
          "factor_name",
          "current_value",
          "current_percentile",
          "history_length_days",
          "percentiles"
        ],
        "title": "FactorAnalysisResponse",
        "type": "object"
      },
      "FactorRequest": {
        "properties": {
          "data_source": {
            "default": "yfinance",
            "enum": [
              "yfinance",
              "vnstock",
              "csv"
            ],
            "title": "Data Source",
            "type": "string"
          },
          "date_range": {
            "$ref": "#/components/schemas/DateRange"
          },
          "factor_type": {
            "enum": [
              "moving_average",
              "distance_from_ma",
              "bollinger",
              "donchian",
              "distance_from_peak"
            ],
            "title": "Factor Type",
            "type": "string"
          },
          "ma_type": {
            "default": "sma",
            "enum": [
              "sma",
              "ema",
              "wma"
            ],
            "title": "Ma Type",
            "type": "string"
          },
          "period": {
            "default": 20,
            "title": "Period",
            "type": "integer"
          },
          "std_dev": {
            "default": 2.0,
            "title": "Std Dev",
            "type": "number"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "date_range",
          "factor_type"
        ],
        "title": "FactorRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PredefinedRarityRequest": {
        "properties": {
          "watchlist_id": {
            "exclusiveMinimum": 0.0,
            "title": "Watchlist Id",
            "type": "integer"
          }
        },
        "required": [
          "watchlist_id"
        ],
        "title": "PredefinedRarityRequest",
        "type": "object"
      },
      "PredefinedRarityResponse": {
        "properties": {
          "available_instruments": {
            "title": "Available Instruments",
            "type": "integer"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "instruments": {
            "items": {
              "$ref": "#/components/schemas/PredefinedRarityInstrumentStatus"
            },
            "title": "Instruments",
            "type": "array"
          },
          "missing_instrument_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Missing Instrument Ids",
            "type": "array"
          },
          "percentile_columns": {
            "items": {
              "type": "string"
            },
            "title": "Percentile Columns",
            "type": "array"
          },
          "requested_instruments": {
            "title": "Requested Instruments",
            "type": "integer"
          },
          "stale_instrument_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Stale Instrument Ids",
            "type": "array"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/PredefinedRarityTable"
            },
            "title": "Tables",
            "type": "array"
          },
          "watchlist_id": {
            "title": "Watchlist Id",
            "type": "integer"
          },
          "watchlist_name": {
            "title": "Watchlist Name",
            "type": "string"
          }
        },
        "required": [
          "watchlist_id",
          "watchlist_name",
          "requested_instruments",
          "available_instruments",
          "stale_instrument_ids",
          "missing_instrument_ids",
          "instruments",
          "percentile_columns",
          "tables"
        ],
        "title": "PredefinedRarityResponse",
        "type": "object"
      },
...
      "RarityRequest": {
        "properties": {
          "factor_type": {
            "enum": [
              "moving_average",
              "distance_from_ma",
              "bollinger",
              "donchian",
              "distance_from_peak"
            ],
            "title": "Factor Type",
            "type": "string"
          },
          "instrument_id": {
            "exclusiveMinimum": 0.0,
            "title": "Instrument Id",
            "type": "integer"
          },
          "ma_type": {
            "default": "sma",
            "enum": [
              "sma",
              "ema",
              "wma"
            ],
            "title": "Ma Type",
            "type": "string"
          },
          "period": {
            "default": 200,
            "title": "Period",
            "type": "integer"
          },
          "quick_recovery_days": {
            "default": 5,
            "title": "Quick Recovery Days",
            "type": "integer"
          },
          "recovery_mode": {
            "default": "price",
            "enum": [
              "price",
              "factor"
            ],
            "title": "Recovery Mode",
            "type": "string"
          },
          "std_dev": {
            "default": 2.0,
            "title": "Std Dev",
            "type": "number"
          },
          "zones": {
            "default": [
              1,
              5,
              10,
              15,
              20,
              25,
              30
            ],
            "items": {
              "type": "integer"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "instrument_id",
          "factor_type"
        ],
        "title": "RarityRequest",
        "type": "object"
      },
...
      "RegimeRequest": {
        "properties": {
          "data_source": {
            "default": "yfinance",
            "enum": [
              "yfinance",
              "vnstock",
              "csv"
            ],
            "title": "Data Source",
            "type": "string"
          },
          "date_range": {
            "$ref": "#/components/schemas/DateRange"
          },
          "factor_type": {
            "enum": [
              "moving_average",
              "distance_from_ma",
              "bollinger",
              "donchian",
              "distance_from_peak"
            ],
            "title": "Factor Type",
            "type": "string"
          },
          "lower_threshold": {
            "default": 0.4,
            "title": "Lower Threshold",
            "type": "number"
          },
          "ma_type": {
            "default": "sma",
            "enum": [
              "sma",
              "ema",
              "wma"
            ],
            "title": "Ma Type",
            "type": "string"
          },
          "period": {
            "default": 20,
            "title": "Period",
            "type": "integer"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "title": "Symbols",
            "type": "array"
          },
          "threshold": {
            "default": 0.0,
            "title": "Threshold",
            "type": "number"
          },
          "upper_threshold": {
            "default": 0.6,
            "title": "Upper Threshold",
            "type": "number"
          }
        },
        "required": [
          "symbols",
          "date_range",
          "factor_type"
        ],
        "title": "RegimeRequest",
        "type": "object"
      },
      "RegimeResponse": {
        "properties": {
          "breadth": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Breadth",
            "type": "object"
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Labels",
            "type": "object"
          }
        },
        "required": [
          "labels",
          "breadth"
        ],
        "title": "RegimeResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backtesting and factor analysis API",
    "title": "Trading Statistics Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/factors/analyze": {
      "post": {
        "operationId": "analyze_factor_endpoint_factors_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactorAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Factor Endpoint",
        "tags": [
          "factors"
        ]
      }
    },
    "/factors/predefined-rarity": {
      "post": {
        "operationId": "predefined_rarity_endpoint_factors_predefined_rarity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredefinedRarityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredefinedRarityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predefined Rarity Endpoint",
        "tags": [
          "factors"
        ]
      }
    },
    "/factors/rarity": {
      "post": {
        "operationId": "rarity_analysis_endpoint_factors_rarity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RarityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RarityAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rarity Analysis Endpoint",
        "tags": [
          "factors"
        ]
      }
    },
    "/factors/regime": {
      "post": {
        "operationId": "detect_regime_endpoint_factors_regime_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegimeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detect Regime Endpoint",
        "tags": [
          "factors"
        ]
      }
    },
    "/factors/universe": {
      "post": {
        "operationId": "analyze_universe_endpoint_factors_universe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrossSectionalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrossSectionalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Universe Endpoint",
        "tags": [
          "factors"
        ]
      }
    },
...
  }
}